requirements.txt
click==6.6
dominate==2.2.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycrypto==2.6.1
pymongo==3.3.0
requests==2.11.1
visitor==0.1.3
Werkzeug==0.11.10
xlwt==1.1.2
xmltodict==0.9.2
